+++ instrumentation/opentelemetry-instrumentation-falcon/tests/app.py
@@ -33,6 +33,18 @@ class ErrorResource:
         print(non_existent_var)  # noqa
 
 
+class CustomResponseHeaderResource:
+    def on_get(self, _, resp):
+        # pylint: disable=no-member
+        resp.status = falcon.HTTP_201
+        resp.set_header("content-type", "text/plain; charset=utf-8")
+        resp.set_header("content-length", "0")
+        resp.set_header(
+            "my-custom-header", "my-custom-value-1,my-custom-header-2"
+        )
+        resp.set_header("dont-capture-me", "test-value")
+
+
 def make_app():
     if hasattr(falcon, "App"):
         # Falcon 3
@@ -43,4 +55,7 @@ def make_app():
     app.add_route("/hello", HelloWorldResource())
     app.add_route("/ping", HelloWorldResource())
     app.add_route("/error", ErrorResource())
+    app.add_route(
+        "/test_custom_response_headers", CustomResponseHeaderResource()
+    )
     return app